# integrations/csm-integration-service/openapi.yaml
openapi: 3.0.1
info:
  title: CSM Integration Service API
  version: 1.0.0
  description: >
    Project/account search and their Contacts sub-resource, incident
    creation, and alert-incident mapping create/lookup, for third-party
    (M2M) consumers. Fronted by Choreo's API Manager gateway (subscription +
    client credentials) — this service performs no inbound auth of its own.
security:
  - oauth2ClientCredentials: []
servers:
  - url: "{server}:{port}/"
    variables:
      server:
        default: http://localhost
      port:
        default: "8080"
paths:
...
  /projects/search:
    post:
      summary: Search projects.
      operationId: searchProjects
      requestBody:
        description: Project search payload
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectSearchPayload'
      responses:
        "200":
          description: Projects matching the search query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSearchResponse'
        "400":
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized (upstream reported an authorization failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    oauth2ClientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          # Placeholder — fill in with the actual token endpoint for this service's
          # Choreo-managed API once it's provisioned. Not a real/guessed URL.
          tokenUrl: "<token_url>"
          scopes: {}
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          description: Number of results per page.
        offset:
          type: integer
          description: Zero-based offset into the result set.
...
    Project:
      type: object
      properties:
        id:
          type: string
        accountId:
          type: string
        sfId:
          type: string
        name:
          type: string
        projectKey:
          type: string
        subscriptionType:
          type: string
          enum:
            - development_support
            - managed_cloud_subscription
            - evaluation_subscription
            - subscription
            - cloud_evaluation_support
            - internal
            - platformer_subscription
            - cloud_support
            - professional_services
        startDate:
          type: string
          format: date-time
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        createdOn:
          type: string
          format: date-time
        updatedOn:
          type: string
          format: date-time
        suspensionProcessState:
          type: object
          additionalProperties: true
          nullable: true
          description: >
            Free-form, per-dimension Account Closure Process (ACP) suspension-process
            state (event type + action results) written by an existing ServiceNow
            suspension flow. This is opaque JSON passed through as-is — its shape is
            not defined or validated by this layer.
    ProjectSearchPayload:
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        searchQuery:
          type: string
          description: Case-insensitive match against name, key, and subscription type.
    ProjectSearchResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
...